Fix view storage losing state on configuration change
