+++ superwall/src/main/java/com/superwall/sdk/Superwall.kt
@@ -3,6 +3,7 @@ package com.superwall.sdk
 import android.app.Application
 import android.content.Context
 import android.net.Uri
+import androidx.lifecycle.ViewModelProvider
 import androidx.work.WorkManager
 import com.superwall.sdk.analytics.internal.track
 import com.superwall.sdk.analytics.internal.trackable.InternalSuperwallEvent
@@ -25,12 +26,23 @@ import com.superwall.sdk.paywall.presentation.internal.dismiss
 import com.superwall.sdk.paywall.presentation.internal.state.PaywallResult
 import com.superwall.sdk.paywall.vc.PaywallViewController
 import com.superwall.sdk.paywall.vc.SuperwallPaywallActivity
+import com.superwall.sdk.paywall.vc.SuperwallStoreOwner
+import com.superwall.sdk.paywall.vc.ViewModelFactory
+import com.superwall.sdk.paywall.vc.ViewStorageViewModel
 import com.superwall.sdk.paywall.vc.delegate.PaywallViewControllerEventDelegate
 import com.superwall.sdk.paywall.vc.web_view.messaging.PaywallWebEvent
-import com.superwall.sdk.paywall.vc.web_view.messaging.PaywallWebEvent.*
+import com.superwall.sdk.paywall.vc.web_view.messaging.PaywallWebEvent.Closed
+import com.superwall.sdk.paywall.vc.web_view.messaging.PaywallWebEvent.Custom
+import com.superwall.sdk.paywall.vc.web_view.messaging.PaywallWebEvent.InitiatePurchase
+import com.superwall.sdk.paywall.vc.web_view.messaging.PaywallWebEvent.InitiateRestore
+import com.superwall.sdk.paywall.vc.web_view.messaging.PaywallWebEvent.OpenedDeepLink
+import com.superwall.sdk.paywall.vc.web_view.messaging.PaywallWebEvent.OpenedURL
+import com.superwall.sdk.paywall.vc.web_view.messaging.PaywallWebEvent.OpenedUrlInSafari
 import com.superwall.sdk.storage.ActiveSubscriptionStatus
 import com.superwall.sdk.store.ExternalNativePurchaseController
-import kotlinx.coroutines.*
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -38,7 +50,8 @@ import kotlinx.coroutines.flow.drop
 import kotlinx.coroutines.flow.filter
 import kotlinx.coroutines.flow.take
 import kotlinx.coroutines.flow.update
-import java.util.*
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 
 class Superwall(
     context: Context,
@@ -54,6 +67,14 @@ class Superwall(
 
     // Add a private variable for the purchase task
     private var purchaseTask: Job? = null
+
+    private val viewStorageViewModel =
+        ViewModelProvider(
+            SuperwallStoreOwner,
+            ViewModelFactory(),
+        ).get(ViewStorageViewModel::class.java)
+
+    internal fun viewStore(): ViewStorageViewModel = viewStorageViewModel
 
     internal val presentationItems: PresentationItems = PresentationItems()
 
@@ -165,7 +186,8 @@ class Superwall(
      */
     val latestPaywallInfo: PaywallInfo?
         get() {
-            val presentedPaywallInfo = dependencyContainer.paywallManager.presentedViewController?.info
+            val presentedPaywallInfo =
+                dependencyContainer.paywallManager.presentedViewController?.info
             return presentedPaywallInfo ?: presentationItems.paywallInfo
         }
 
@@ -305,7 +327,8 @@ class Superwall(
                 )
         }
 
-        val cachedSubsStatus = dependencyContainer.storage.get(ActiveSubscriptionStatus) ?: SubscriptionStatus.UNKNOWN
+        val cachedSubsStatus =
+            dependencyContainer.storage.get(ActiveSubscriptionStatus) ?: SubscriptionStatus.UNKNOWN
         setSubscriptionStatus(cachedSubsStatus)
 
         addListeners()
@@ -350,7 +373,8 @@ class Superwall(
      */
     fun togglePaywallSpinner(isHidden: Boolean) {
         ioScope.launch {
-            val paywallViewController = dependencyContainer.paywallManager.presentedViewController ?: return@launch
+            val paywallViewController =
+                dependencyContainer.paywallManager.presentedViewController ?: return@launch
             paywallViewController.togglePaywallSpinner(isHidden)
         }
     }
@@ -374,7 +398,9 @@ class Superwall(
      * Removes all of Superwall's pending local notifications.
      */
     fun cancelAllScheduledNotifications() {
-        WorkManager.getInstance(context).cancelAllWorkByTag(SuperwallPaywallActivity.NOTIFICATION_CHANNEL_ID)
+        WorkManager
+            .getInstance(context)
+            .cancelAllWorkByTag(SuperwallPaywallActivity.NOTIFICATION_CHANNEL_ID)
     }
 
     // MARK: - Reset
@@ -478,6 +504,7 @@ class Superwall(
                         closeReason = PaywallCloseReason.ManualClose,
                     )
                 }
+
                 is InitiatePurchase -> {
                     if (purchaseTask != null) {
                         // If a purchase is already in progress, do not start another
@@ -496,18 +523,23 @@ class Superwall(
                             }
                         }
                 }
+
                 is InitiateRestore -> {
                     dependencyContainer.transactionManager.tryToRestore(paywallViewController)
                 }
+
                 is OpenedURL -> {
                     dependencyContainer.delegateAdapter.paywallWillOpenURL(url = paywallEvent.url)
                 }
+
                 is OpenedUrlInSafari -> {
                     dependencyContainer.delegateAdapter.paywallWillOpenURL(url = paywallEvent.url)
                 }
+
                 is OpenedDeepLink -> {
                     dependencyContainer.delegateAdapter.paywallWillOpenDeepLink(url = paywallEvent.url)
                 }
+
                 is Custom -> {
                     dependencyContainer.delegateAdapter.handleCustomPaywallAction(name = paywallEvent.string)
                 }
