+++ superwall/src/main/java/com/superwall/sdk/paywall/vc/SuperwallStoreOwner.kt
@@ -1,0 +1,21 @@
+package com.superwall.sdk.paywall.vc
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.ViewModelStore
+import androidx.lifecycle.ViewModelStoreOwner
+import com.superwall.sdk.paywall.vgc.ViewStorageViewModel
+
+internal object SuperwallStoreOwner : ViewModelStoreOwner {
+    override val viewModelStore: ViewModelStore = ViewModelStore()
+}
+
+internal class ViewModelFactory : ViewModelProvider.Factory {
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        if (modelClass.isAssignableFrom(ViewStorageViewModel::class.java)) {
+            @Suppress("UNCHECKED_CAST")
+            return ViewStorageViewModel() as T
+        }
+        throw IllegalArgumentException("Unknown ViewModel class")
+    }
+}
